apps/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
langgraph==0.1.5
langchain==0.2.5
langchain-community==0.2.5
langchain-openai==0.1.8
litellm==1.40.10
lancedb==0.8.2
pyarrow==14.0.2
numpy==1.26.4
sentence-transformers==3.0.0
pydantic==2.7.3
pydantic-settings==2.3.1
apscheduler==3.10.4
aiofiles==23.2.1
aiohttp==3.9.5
httpx==0.27.0
python-dotenv==1.0.1
watchfiles==0.22.0
PyGithub==2.3.0
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.133.0
cryptography==42.0.8
keyring==25.2.1
tiktoken==0.7.0
tenacity==8.3.0
rich==13.7.1
